Iodine: Handle when Alice goes faster then slower than Client
Handle the case when Alice goes faster (adds packets to the TUN) and
then slower (the packet sends take less time than Alice pacing) than the
weird client.

This is done with an attribute that tracks the number of packets that
have been admitted.

All packets are present in Alice from the start.  However, not all may
be transmitted.  numAdmittedPkts tracks how many packets are eligible to
be sent to Bob.  When Alice sends a packet, that number is decremented
by 1 (is at least 0).  When the packet timeout expires, numAdmittedPkts
is incremented.  Alice can only send when numAdmittedPkts is not 0.