Generation: Allow generating files without a shadow file
Allow model generation even a shadow file is not passed.  That means,
generation should use the Maude_HCS names, link characteristics (or
defaults), etc.

Continue to translate names to shadow names when available.

Set the zone files to the pwnd2 domain name, not pwnd host name.

The code could be cleaned up further, but need to attend to other
issues.