Iodine: Implement last frag delay at server
Implement delaying the response to the last fragment of a transmission.

The server sends a message to itself with a fixed delay (defined in
parameters).  It then proceeds to send the message, which includes the
response to send.

Note: This has not been fully tested yet---not clear what happens if
another query is received in the intermediate.