Prob Iodine: Let Maude Alice create its list of packets to send
If queuePopulated is set to true, then Alice will use a queue list
provided in queue.

Otherwise, it will use the file size and min and max packet sizes (in
bytes) to create a list of packets to send.  It does so at start time,
before it begins sending packets.