Mastodon: Fix defects
Fix issues with Maude not recognizing HttpMsg as a Msg or (to Addr from
Addr' : HttpRequest) type messages. (This fix is to add more eqs)

Store media with their URL in the media Map (otherwise, attachments
cannot be downloaded).

Handle requests for posting and getting at server.

Add unit tests.