Mastodon: Add Ids with urls in toots
Add list of attachments ids with the list of attachment URLs.  Decided
to juxtapose the two instead of zipping them because it would have
caused lots of big changes, and we may want to add more information with
the toot.
Ids are required to recombine the payloads of images _in order_.