Mastodon: Model files by more than their filename
Model the files by their filename AND size and payload.

To allow making progress on Mastodon, we did not specify a file beyond a
simple filename.  This commit enriches the file with a size (in Bytes)
and a payload (a byte field).

This has lead to a lot of changes, but they are simplifying the code a
lot.