Common: Add Alice and Bob actors
Add Alice and Bob actors.  These are different than the actors in
Iodine, which can only communicate with Iodine client or server.

These Alice and Bob sit on top and can communicate with Raceboat, and
the Iodine send app.

Alice and Bob send each other files; Alice sends a file to Iodine.